Method to compute ingress path external of inbound mail qualifier
