+++ src/backend/main/py.pi/caliopen_pi/features/mail.py
@@ -2,12 +2,12 @@
 """Caliopen mail message privacy features extraction methods."""
 from __future__ import absolute_import, print_function, unicode_literals
 
-import re
 import logging
 
 import pgpy
 
 from .spam import SpamScorer
+from .ingress_path import get_ingress_features
 from ..parameters import PIParameter
 
 log = logging.getLogger(__name__)
@@ -90,7 +90,7 @@ class InboundMailFeature(object):
     @property
     def spam_informations(self):
         """Return a global spam_score and related features."""
-        spam = SpamScorer(self.message.mail.headers())
+        spam = SpamScorer(self.message.mail)
         return {'spam_score': spam.score,
                 'spam_method': spam.method,
                 'is_spam': spam.is_spam}
@@ -118,83 +118,11 @@ class InboundMailFeature(object):
         is_encrypted = True if encrypted_parts else False
         return {'message_encrypted': is_encrypted}
 
-    def get_ingress_features(self):
-        """Try to find information about ingress server that send this mail."""
-        def get_host(line):
-            if ' ' not in line:
-                return None
-            parts = line.split(' ')
-            return parts[0]
-
-        def parse_ingress(header):
-            header = header.replace('\n', '')
-            search = re.compile(r'.*using (\S+) with cipher ([\S-]+)',
-                                re.MULTILINE)
-            match = search.match(header)
-            if match:
-                return match.groups()
-            return None
-
-        found_features = {}
-        received = self.message.headers.get('Received')
-        if not received:
-            return {}
-
-        # First step Search for 'paths' (from / by) information
-        search = re.compile(r'^from (.*) by (.*)', re.MULTILINE + re.DOTALL)
-        paths = []
-        for r in received:
-            r = r.replace('\n', '')
-            match = search.match(r)
-            if match:
-                groups = match.groups()
-                from_ = get_host(groups[0])
-                by = get_host(groups[1])
-                if from_ and by:
-                    paths.append((from_, by, groups))
-                else:
-                    log.warn('Invalid from {} or by {} path in {}'.
-                             format(from_, by, r))
-            else:
-                if r.startswith('by'):
-                    # XXX first hop, to consider ?
-                    pass
-                else:
-                    log.warn('Received header, does not match format {}'.
-                             format(r))
-
-        # Second step: qualify path if internal and try to find the ingress one
-        ingress = None
-        internal_hops = 0
-        for path in paths:
-            is_internal = False
-            for internal in self.internal_domains:
-                if internal in path[0]:
-                    is_internal = True
-                else:
-                    if internal in path[1] and internal not in path[0]:
-                        ingress = path
-                        break
-            if is_internal:
-                internal_hops += 1
-
-        # Qualify ingress connection
-        if ingress:
-            cnx_info = parse_ingress(ingress[2][0])
-            if cnx_info and len(cnx_info) > 1:
-                found_features.update({'ingress_socket_version': cnx_info[0],
-                                       'ingress_cipher': cnx_info[1]})
-            found_features.update({'ingress_server': ingress[0]})
-
-        # Try to count external hops
-        external_hops = len(paths) - internal_hops
-        found_features.update({'nb_external_hops': external_hops})
-        return found_features
-
     def process(self):
         """Process the message for privacy features extraction."""
         features = self._features.copy()
-        features.update(self.get_ingress_features())
+        received = self.message.headers.get('Received', [])
+        features.update(get_ingress_features(received, self.internal_domains))
         mx = features.get('ingress_server')
         reputation = None if not mx else self.emitter_reputation(mx)
         features['mail_emitter_mx_reputation'] = reputation
